+++ examples/typescript/swap.ts
@@ -23,7 +23,7 @@ import {
   Network,
   NetworkToNetworkName,
   U64,
-  ViewRequestData,
+  InputViewRequestData,
 } from "@aptos-labs/ts-sdk";
 import { createInterface } from "readline";
 // Default to devnet, but allow for overriding
@@ -40,7 +40,7 @@ const getOptimalLpAmount = async (
   token1Addr: AccountAddress,
   token2Addr: AccountAddress,
 ): Promise<void> => {
-  const payload: ViewRequestData = {
+  const payload: InputViewRequestData = {
     function: `${swap.toString()}::router::optimal_liquidity_amounts`,
     functionArguments: [token1Addr.toString(), token2Addr.toString(), false, "200000", "300000", "200", "300"],
   };